#48: Move all activities into activites folder
