+++ ShinyExpenseTracker/AndroidManifest.xml
@@ -35,15 +35,15 @@
             android:label="@string/title_activity_view_expense_claim" >
         </activity>
         <activity
-            android:name="ca.ualberta.cs.shinyexpensetracker.AddExpenseClaimActivity"
+            android:name="ca.ualberta.cs.shinyexpensetracker.activities.AddExpenseClaimActivity"
             android:label="@string/title_activity_add_expense_claim" >
         </activity>
         <activity
-            android:name="ca.ualberta.cs.shinyexpensetracker.ExpenseItemActivity"
+            android:name="ca.ualberta.cs.shinyexpensetracker.activities.ExpenseItemActivity"
             android:label="@string/title_activity_expense_item" >
         </activity>
         <activity
-            android:name="ca.ualberta.cs.shinyexpensetracker.AddDestinationActivity"
+            android:name="ca.ualberta.cs.shinyexpensetracker.activities.AddDestinationActivityionActivity"
             android:label="@string/title_activity_add_destination" >
         </activity>
         <activity
@@ -55,7 +55,7 @@
             android:label="@string/title_activity_claim_summary" >
         </activity>
 		<activity
-            android:name=".ExpenseItemDetailView"
+            android:name="ca.ualberta.cs.shinyexpensetracker.activities.ExpenseItemDetailView"
             android:label="@string/title_activity_expense_item_detail_view" >
         </activity>
     </application>
