+++ ShinyExpenseTracker/res/layout/activity_add_expense_claim.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="ca.ualberta.cs.shinyexpensetracker.AddExpenseClaimActivity" >
+    tools:context="ca.ualberta.cs.shinyexpensetracker.activities.AddExpenseClaimActivity" >
 
     <EditText
         android:id="@+id/editTextExpenseClaimName"
