+++ ShinyExpenseTracker/res/layout/activity_expense_item_detail_view.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="ca.ualberta.cs.shinyexpensetracker.ExpenseItemDetailView" >
+    tools:context="ca.ualberta.cs.shinyexpensetracker.activities.ExpenseItemDetailView" >
 
     <TextView
         android:id="@+id/expenseItemCategoryTitle"
