+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/activities/TabbedSummaryActivity.java
@@ -11,9 +11,6 @@ import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
 import android.view.Menu;
 import android.view.MenuItem;
-import ca.ualberta.cs.shinyexpensetracker.AddDestinationActivity;
-import ca.ualberta.cs.shinyexpensetracker.AddExpenseClaimActivity;
-import ca.ualberta.cs.shinyexpensetracker.ExpenseItemActivity;
 import ca.ualberta.cs.shinyexpensetracker.R;
 import ca.ualberta.cs.shinyexpensetracker.SectionsPagerAdapter;
 import ca.ualberta.cs.shinyexpensetracker.fragments.ClaimSummaryFragment;
